+++ app/src/main/java/com/navapp/navigation/util/math/LocationCluster.java
@@ -1,11 +1,11 @@
 package com.navapp.navigation.util.math;
 
+import com.navapp.navigation.util.math.TSP.TSPSolver;
+
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Deque;
 import java.util.List;
 
 public class LocationCluster implements MathLocation {
